+++ src/widgets/logwidget.cpp
@@ -103,9 +103,10 @@ void LogWidget::storeSettings() const {
  * Adds a log entry
  */
 void LogWidget::log(LogWidget::LogType logType, QString text) {
-    // ignore libpng sRGB profile warnings
-    if (logType == WarningLogType && text.contains(
-            "libpng warning: iCCP: known incorrect sRGB profile")) {
+    // ignore "libpng sRGB profile" and "QXcbConnection: XCB error: 8" warnings
+    if (logType == WarningLogType &&
+        (text.contains("libpng warning: iCCP: known incorrect sRGB profile") ||
+         text.contains("QXcbConnection: XCB error: 8"))) {
         return;
     }