+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - you can click this button to enable note editing again
 - double-clicking on a tag now assigns the tag to the current note
   (for [#896](https://github.com/pbek/QOwnNotes/issues/896))
+- `QXcbConnection: XCB error: 8` warnings are now suppressed
 
 ## 18.04.4
 - fixed a warning if you are un-indenting text at the start or end of the note